+++ scaleph-ui-react/src/pages/Project/Workspace/DataIntegration/SeaTunnel/Dag/components/node/steps/common/schema/fields.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import {InfoCircleOutlined} from "@ant-design/icons";
 import {ProFormGroup, ProFormList, ProFormText} from "@ant-design/pro-components";
 import {getIntl, getLocale} from "@umijs/max";
-import {SchemaParams} from "./constant";
+import {SchemaParams} from "../../constant";
 
 const FieldItem: React.FC = () => {
   const intl = getIntl(getLocale());
