+++ scaleph-ui-react/src/pages/Project/Workspace/DataIntegration/SeaTunnel/Dag/components/node/steps/sink/sink-ossjindo-file-step.tsx
@@ -6,7 +6,7 @@ import {Node, XFlow} from '@antv/xflow';
 import {ModalFormProps} from '@/typings';
 import {STEP_ATTR_TYPE} from '../constant';
 import DataSourceItem from "../dataSource";
-import FileSinkItem from "../fileSink";
+import FileSinkItem from "../common/file/fileSink";
 
 const SinkOSSJindoFileStepForm: React.FC<ModalFormProps<Node>> = ({data, visible, onVisibleChange, onOK}) => {
   const intl = getIntl(getLocale());
