+++ scaleph-ui-react/src/pages/Project/Workspace/DataIntegration/SeaTunnel/Dag/components/node/steps/source/source-ftp-file-step.tsx
@@ -7,7 +7,7 @@ import {ModalFormProps} from '@/typings';
 import {STEP_ATTR_TYPE} from '../constant';
 import {StepSchemaService} from '../helper';
 import DataSourceItem from "../dataSource";
-import FileSourceItem from "../fileSource";
+import FileSourceItem from "../common/file/fileSource";
 
 const SourceFtpFileStepForm: React.FC<ModalFormProps<Node>> = ({data, visible, onVisibleChange, onOK}) => {
     const intl = getIntl(getLocale());
